feat(cohort): roadmap A — node poke, header/calibration fixes, motion meter, cohort starter kit
sensing-server:
- Fix parse_esp32_frame offsets (n_subcarriers u16 @6, seq @12-15, rssi @16, noise @17) in main.rs and csi.rs; per-node RSSI is now real
- Fix calibration feed gate (accept Uncalibrated) and resample mismatched frame lengths; room calibration now reaches 12000 frames (~5 min, 2 nodes)
- Add node_poke task + --node-poke-hz (default 20): 1-byte UDP to each active node so the AP keeps generating CSI-bearing frames (fixes stale nodes on quiet networks)
- 312 tests passed; release binary rebuilt into babybreath-app/bin (gitignored build artifact)

baby UI:
- Remove presence flag / person count (uninformative: true 96% and ~22 persons in an empty room)
- Add Teach-the-room motion meter vs per-room empty baseline; apnea/sleep gate on motion state

cohort/: pip-only Claude-Code-driven starter kit (babybreath.py detect/flash/provision/bootlog/serve/check/walktest/calibrate with PASS/FAIL checkpoints), RUNBOOK.md, CLAUDE.md for the family's Claude, README.md

Docs: CHANGELOG + app README (mesh --filter-mac, placement rule, ESP-IDF-free path, known issues)

+++ rust-port/wifi-densepose-rs/crates/wifi-densepose-sensing-server/src/types.rs
@@ -64,7 +64,7 @@ pub struct Esp32Frame {
     pub magic: u32,
     pub node_id: u8,
     pub n_antennas: u8,
-    pub n_subcarriers: u8,
+    pub n_subcarriers: u16,
     pub freq_mhz: u16,
     pub sequence: u32,
     pub rssi: i8,
